Get basic syntax highlighting working
This pulls the queries and injections straight from the work done for
Helix in [this PR][1], from @zetashift. These should definitely be
upstreamed when possible! They also show up a few bugs in the grammar
as it stands, e.g. keywords get highlighted in the middle of
documentation blocks, where they should be ignored. Still: a start!

[1]: https://github.com/helix-editor/helix/pull/7724