Restructure the repo/package layout
Make the extension itself a child directory within the repo, not the
root of the repo. This keeps the distinction between things which are
part of the distributed artifact and things which simply exist to help
build it stronger, and will get rid of some annoyances around pnpm and
symlinks when trying to distribute it!